python/msgraph/requirements.txt
--index-url https://pypi.python.org/simple
azure-identity==1.3.1
--index-url https://test.pypi.org/simple
msgraphcore==0.0.2